写信号数据生成测试类,能随机的根据规则和已有车辆生成信号数据,并判断预警接口的结果是否正确,输出每个题目的解答情况日志. 写对公式安全检查的工具类. WarnVo类的warnLevel由Byte改为String方便显示"不报警",原先用null表示不报警.

+++ src/main/java/com/xiaomi/service/impl/RuleServiceImpl.java
@@ -18,6 +18,7 @@ import com.xiaomi.domain.vo.RuleVo;
 import com.xiaomi.exception.DataNotExistException;
 import com.xiaomi.mapper.RuleMapper;
 import com.xiaomi.service.RuleService;
+import com.xiaomi.util.SafeFormulaCheck;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.StringRedisTemplate;
@@ -94,6 +95,11 @@ public class RuleServiceImpl extends ServiceImpl<RuleMapper, Rule> implements Ru
     @Override
     public void insertRule(RuleDto ruleDto) {
         Rule rule = BeanUtil.copyProperties(ruleDto, Rule.class);
+        String formula = rule.getFormulaRateConfig().getFormula(); // 形如"10 * ( ${Mx} - ${Mi} ) / ${Mx}"
+        // 对公式进行安全检查,防止注入攻击
+        if (!SafeFormulaCheck.isSafeFormula(formula)) {
+            throw new IllegalArgumentException("不安全的公式:" + formula);
+        }
         save(rule);
         // 删除与这个规则编号warnId相关的查询缓存
         deleteRelatedRuleCache(rule.getWarnId(), rule.getBatteryType());
